fix(encoding): move encode and decode num to script helper
