+++ lib/script_helper.rb
@@ -1,0 +1,13 @@
+module ScriptHelper
+  def encode_num(num)
+    return '' if num.zero?
+
+    int_to_little_endian(num, 1)
+  end
+
+  def decode_num(string)
+    return 0 if string.empty?
+
+    little_endian_to_int(string)
+  end
+end
